statstool: add some more options
Add -X option to allow us to scale the current for use by a
current meter clamp (which requires us to scale by 1000.0).

Rename -s option -t a, to run stats on all samples from a log;
this is useful on non-tagged data samples.

Add -s to dump out a summary (used to do this by default).

Add -S to dump out stats at the end of each tests.

Make -sS on by default, unless either is supplied, in which case
one has to supply the other option to turn both on again.

Signed-off-by: Colin Ian King <colin.king@canonical.com>